refactor: move language code to a separate file
