+++ index.html
@@ -162,16 +162,17 @@
         <span class="footer__lang" id="en">en</span>
       </li>
       <li class="footer__nav-item">
-        <a class="footer__nav-link"href="https://github.com/Stesho">Github</a>
+        <a class="footer__nav-link"href="https://github.com/Stesho" target="_blank">Github</a>
       </li>
       <li class="footer__nav-item">
         <p>2022</p>
       </li>
       <li class="footer__nav-item">
-        <a href="https://rs.school/js/"><img src="assets/images/rs_school.svg" width="75px" alt="rs school logo"></a>
+        <a href="https://rs.school/js/" target="_blank"><img src="assets/images/rs_school.svg" width="75px" alt="rs school logo"></a>
       </li>
     </ul>
   </footer>
+  <script type="module" src="./assets/js/language.js"></script>
   <script type="module" src="./assets/js/script.js"></script>
 </body>
 </html>
